fix #144 #149 #152 test_case4 and path search and returned values from set_redirection
