+++ includes/minishell_tnishina.h
@@ -19,6 +19,8 @@
 
 # define PROMPT "minishell$ "
 # define EXIT_PROMPT "exit\n"
+# define STATUS_CANNOT_EXECUTE 126
+# define STATUS_COMMAND_NOT_FOUND 127
 # define STATUS_SYNTAX_ERR 258
 # define REDIRECT_IN "<"
 # define REDIRECT_OUT ">"
@@ -44,6 +46,8 @@
 # define SYNTAX_ERROR_MSG "syntax error near unexpected token"
 # define FD_OOR_MSG "file descriptor out of range"
 # define FD_ERROR_MSG "Bad file descriptor"
+# define IS_DIR_ERROR_MSG "is a directory"
+# define COMMAND_NOT_FOUND_ERR_MSG "command not found"
 
 /*
 ** macro
